restrict access to resourcess to their owner only

+++ src/api/services/FileService.ts
@@ -5,13 +5,16 @@ import { Database } from 'src/database/db';
 import { Prisma } from "@prisma/client";
 import { getPagination } from '../utils';
 import { Pagination } from '../types';
+import { UserService } from './UserService';
+import { NotFoundError } from 'routing-controllers';
+import { AccessDeniedError } from '../errors/AccessDeniedError';
 
 export const prisma = new PrismaClient();
 
 @Service()
 export class FileService {
 
-    constructor(private readonly s3Service: S3Service, private readonly database: Database) { }
+    constructor(private readonly s3Service: S3Service, private readonly userService: UserService, private readonly database: Database) { }
 
     public async generatePutPreSignedUrl(userId: string) {
         return this.s3Service.generatePutPreSignedUrl({ userId });
@@ -21,9 +24,12 @@ export class FileService {
         return this.database.files.create(data);
     }
 
-    public findFiles(pagination?: Pagination) {
+    public findFiles(ownerId: string, pagination?: Pagination) {
         const prismaPagination = getPagination(pagination);
         return this.database.files.find({
+            where: {
+                ownerId
+            },
             take: prismaPagination.take,
             skip: prismaPagination.skip
         });
@@ -33,8 +39,24 @@ export class FileService {
         return this.database.files.deleteAll();
     }
 
-    public deleteOne(id: string) {
-        this.s3Service.deleteFile(id);
+    public async deleteOne(id: string, ownerId: string) {
+        const user = await this.userService.getUser(ownerId);
+
+        if (!user) {
+            throw new NotFoundError('User not found');
+        }
+
+        const file = await this.database.files.findUnique({ where: { id } });
+
+        if (!file) {
+            throw new NotFoundError('File not found');
+        }
+
+        if (file.ownerId !== user.id) {
+            throw new AccessDeniedError('You have no access to this file');
+        }
+
+        await this.s3Service.deleteFile(id);
         return this.database.files.deleteOne(id);
     }
 }
